+++ src/components/NameValueList/index.js
@@ -1,0 +1,3 @@
+import * as NameValueList from './NameValueList';
+
+export { NameValueList };