+++ src/components/Tile/index.js
@@ -1,0 +1,3 @@
+import Tile from './Tile';
+
+export { Tile };